% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkpop.R
\name{hkpop}
\alias{hkpop}
\title{Get Hong Kong Public Opinion Data.}
\usage{
hkpop(url)
}
\arguments{
\item{url}{the url of a particular survey}
}
\value{
a data frame of the survey table which can be worked with in R
}
\description{
This function allows the user to get Hong Kong public opinion data from the
Public Opinion Programme (POP) survey center at the University of Hong Kong.
}
\details{
The function asks the user to enter the url of a particular survey from the
POP website: \url{https://www.hkupop.hku.hk/english/popexpress/index.html}.
It works on selected surveys: Popularity of Chief Executive (half-yearly only);
Popularity of the Government; Appraisal of Current Conditions; Social Indicators;
Freedom Indicators; Rule of Law Indicators; Media Performance; People's Ethnic Identity;
some National Issues.
}
\examples{
hkpop("https://www.hkupop.hku.hk/english/popexpress/sargperf/sarg/halfyr/datatables.html")
}
